handle Errors

+++ server.cpp
@@ -182,39 +182,46 @@ void server::Run() {
 
 					}
 
-                    if (this->clients[i].login == 2){
+                    else if (this->clients[i].login == 2){
                         if (command == "pwd"){
                             response = ("257: " + this->clients[i].dir + "").c_str() ;
                         }
 
-                        if (command == "mkd"){
-                            string dir_path = args[0];
-                            if(mkdir((this->clients[i].dir + dir_path).c_str(),0777) == 0){
-                                if(this->clients[i].dir.back() != '/')
-                                    this->clients[i].dir.back() = '/';
-                                this->clients[i].dir += dir_path;
-                                response = ("257: " + this->clients[i].dir + " created.").c_str();
+                        else if (command == "mkd"){
+                            if (args.size() != 1)
+                                response = "501: Syntax error in parameters or arguments.";
+                            else{
+                                string dir_path = args[0];
+                                if(mkdir((this->clients[i].dir + dir_path).c_str(),0777) == 0){
+                                    if(this->clients[i].dir.back() != '/')
+                                        this->clients[i].dir += '/';
+                                    this->clients[i].dir += dir_path;
+                                    response = ("257: " + this->clients[i].dir + " created.").c_str();
+                                }
                             }
                         }
 
-                        if (command == "dele"){
-                            string dele_mode = args[0];
-                            if (dele_mode == "-d"){
-                                string dir_path= args[1];
-                                if(rmdir(dir_path.c_str()) == 0)
-                                    response = ("250: " + dir_path + " deleted.").c_str();
-                            }
+                        else if (command == "dele"){
+                            if (args.size() != 2)
+                                response = "501: Syntax error in parameters or arguments.";
+                            else{
+                                string dele_mode = args[0];
+                                if (dele_mode == "-d"){
+                                    string dir_path= args[1];
+                                    if(rmdir(dir_path.c_str()) == 0)
+                                        response = ("250: " + dir_path + " deleted.").c_str();
+                                }
 
-                            //Have a problem
-                            else if (dele_mode == "-f"){
-                                string filename= args[1];
-                                if(remove(filename.c_str()) == 0 )
-                                    response = ("250: " + filename + " deleted.").c_str();
+                                //Have a problem
+                                else if (dele_mode == "-f"){
+                                    string filename= args[1];
+                                    if(remove(filename.c_str()) == 0 )
+                                        response = ("250: " + filename + " deleted.").c_str();
+                                }
                             }
-                            
                         }
 
-                        if (command == "cwd"){
+                        else if (command == "cwd"){
                             if (args.size() != 1)
                                 this->clients[i].dir = "./";
                             else{
@@ -231,7 +238,7 @@ void server::Run() {
                         }
                         
                         //Have a problem
-                        if (command == "rename"){
+                        else if (command == "rename"){
                             if (args.size() != 2)
                                 response = "501: Syntax error in parameters or arguments.";
                             else{
@@ -242,7 +249,8 @@ void server::Run() {
                             }
                         }
 
-                        if (command == "ls"){
+                        //??????????????
+                        else if (command == "ls"){
                             struct dirent *entry;
                             DIR *dir = opendir(this->clients[i].dir.c_str());
 
@@ -257,17 +265,18 @@ void server::Run() {
                             
                             response = "226: List transfer done.";
                         }
+
+                        else response = "501: Syntax error in parameters or arguments.";
                         
                     
                     }
-
+                    else
+                        response = "332: Need acount for login.";
                     //send responce to client
                     
                     char* message = &response[0];
                     if(send(sd, message, strlen(message), 0) != strlen(message))
                         cerr << ("send() sent a different number of bytes than expected");
-                    if(command == "ls")
-                        this->createDataChannel(data);
                 }
             }
         }
@@ -284,43 +293,4 @@ void server::Run() {
 		}
 		cout << endl;*/
 	}
-}
-
-
-void server::createDataChannel(string data){
-    cout << "here\n";
-    int server_fd, new_socket, valread;
-    struct sockaddr_in address;
-    int opt = 1;
-    int addrlen = sizeof(address);
-    char buffer[BUFSIZE] = {0};
-       
-    if ((server_fd = socket(AF_INET, SOCK_STREAM, 0)) == 0){
-        perror("socket failed");
-        exit(EXIT_FAILURE);
-    }
-       
-    if (setsockopt(server_fd, SOL_SOCKET, SO_REUSEADDR | SO_REUSEPORT, &opt, sizeof(opt))){
-        perror("setsockopt");
-        exit(EXIT_FAILURE);
-    }
-    address.sin_family = AF_INET;
-    address.sin_addr.s_addr = INADDR_ANY;
-    address.sin_port = htons(this->dataChannelPort);
-       
-    if (bind(server_fd, (struct sockaddr *)&address, sizeof(address))<0){
-        perror("bind failed");
-        exit(EXIT_FAILURE);
-    }
-    if (listen(server_fd, 3) < 0){
-        perror("listen");
-        exit(EXIT_FAILURE);
-    }
-    if ((new_socket = accept(server_fd, (struct sockaddr *)&address, (socklen_t*)&addrlen))<0){
-        perror("accept");
-        exit(EXIT_FAILURE);
-    }
-    cout << "Asdfd\n";
-    char* message = &data[0];
-    send(new_socket , message , strlen(message) , 0 );
 }
